requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.23.0
lightgbm>=3.3.0
scikit-learn>=1.1.0
openpyxl>=3.0.0
pyyaml>=6.0

# Web scraping
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
selenium>=4.0.0

# Machine learning & optimization
optuna>=3.0.0
shap>=0.41.0
joblib>=1.2.0

# Visualization (optional)
matplotlib>=3.5.0
seaborn>=0.12.0

# Development tools
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.990